pyproject.toml
[project]
name = "comani"
version = "0.1.0"
description = "Personal ComfyUI resource pack for anime-style image generation with VastAI integration"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "delva" },
    { name = "hy1004" }
]
keywords = ["comfyui", "anime", "stable-diffusion", "sdxl", "vastai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "questionary>=2.0.0",
    "paramiko>=4.0.0",
    "pydantic-settings>=2.12.0",
    "grok-api",
    "python-dotenv>=1.2.1",
    "websocket-client>=1.7.0",
]

[project.optional-dependencies]
dev = ["grok-api"]

[project.urls]
Repository = "https://github.com/delva/comani"

[project.scripts]
comani = "comani.cli.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["comani"]

[tool.hatch.build.targets.wheel.sources]
"comani" = "comani"

[tool.hatch.build.targets.sdist]
include = [
    "comani/**",
    "scripts/**",
    "workflows/**",
    "tools/**",
    "README.md",
    "pyproject.toml",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "ruff>=0.8.0",
]

[tool.uv.sources]
grok-api = [
  { path = "../Grok-Api", editable = true, marker = "extra == 'dev'" },
  { git = "https://github.com/Delva0/Grok-Api.git", marker = "extra != 'dev'" },
]
